Network-on-Chip Routing algorithem
Abstract— In recent decades, computer systems have been advanced from relatively simple single-core CISC and RISC architectures to more complex multi-core system-on-chip designs with higher inter-core communication requirements. Networkon- chip architectures emerged as promising solutions for future system-on-chip communication architecture designs. However, the switching and routing algorithm design of network-on-chip communication architectures are still facing great challenges. To address the deficiencies of the existing routing algorithms, a new deterministic network-on-chip routing algorithm-- Edge Traffic Distribution routing and its specifics are proposed in this paper. In the proposed algorithm, the traffic which involves the Edge Routers will be directed to go through the edges first instead of entering into the center of the network. It will greatly alleviate the congestion around the center of the network so as to help improve the overall system performance. Simulation of the proposed Edge Traffic Distribution algorithm and three existing 
 routing algorithms has been carried out using NIRGAM network-on-chip simulator. Simulation results illustrate that the performance of the Edge Traffic Distribution routing algorithm exceeds the performance of the existing routing algorithms in term of average latency and network power consumption.